--- common/src/main/java/com/pedro/common/frame/MediaFrameType.kt
@@ -1,2 +1,0 @@
-package com.pedro.common.frame
-
