+++ rtmp/src/main/java/com/pedro/rtmp/rtmp/RtmpSender.kt
@@ -18,12 +18,11 @@ package com.pedro.rtmp.rtmp
 
 import android.util.Log
 import com.pedro.common.AudioCodec
-import com.pedro.common.BitrateManager
 import com.pedro.common.ConnectChecker
 import com.pedro.common.VideoCodec
+import com.pedro.common.base.BaseSender
 import com.pedro.common.frame.MediaFrame
 import com.pedro.common.onMainThread
-import com.pedro.common.trySend
 import com.pedro.common.validMessage
 import com.pedro.rtmp.flv.BasePacket
 import com.pedro.rtmp.flv.FlvPacket
@@ -34,53 +33,28 @@ import com.pedro.rtmp.flv.video.packet.Av1Packet
 import com.pedro.rtmp.flv.video.packet.H264Packet
 import com.pedro.rtmp.flv.video.packet.H265Packet
 import com.pedro.rtmp.utils.socket.RtmpSocket
-import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.async
-import kotlinx.coroutines.cancelAndJoin
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.isActive
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.runInterruptible
+import kotlinx.coroutines.withContext
 import java.nio.ByteBuffer
-import java.util.concurrent.BlockingQueue
-import java.util.concurrent.LinkedBlockingQueue
 import java.util.concurrent.TimeUnit
 
 /**
  * Created by pedro on 8/04/21.
  */
 class RtmpSender(
-  private val connectChecker: ConnectChecker,
+  connectChecker: ConnectChecker,
   private val commandsManager: CommandsManager
-) {
+): BaseSender(connectChecker, "RtmpSender") {
 
   private var audioPacket: BasePacket = AacPacket()
   private var videoPacket: BasePacket = H264Packet()
-  @Volatile
-  private var running = false
-  private var cacheSize = 200
-
-  private var job: Job? = null
-  private val scope = CoroutineScope(Dispatchers.IO)
-  @Volatile
-  private var queue: BlockingQueue<MediaFrame> = LinkedBlockingQueue(cacheSize)
-  private var audioFramesSent: Long = 0
-  private var videoFramesSent: Long = 0
   var socket: RtmpSocket? = null
-  var droppedAudioFrames: Long = 0
-    private set
-  var droppedVideoFrames: Long = 0
-    private set
-  private val bitrateManager: BitrateManager = BitrateManager(connectChecker)
-  private var isEnableLogs = true
 
-  companion object {
-    private const val TAG = "RtmpSender"
-  }
-
-  fun setVideoInfo(sps: ByteBuffer, pps: ByteBuffer?, vps: ByteBuffer?) {
+  override fun setVideoInfo(sps: ByteBuffer, pps: ByteBuffer?, vps: ByteBuffer?) {
     videoPacket = when (commandsManager.videoCodec) {
       VideoCodec.H265 -> {
         if (vps == null || pps == null) throw IllegalArgumentException("pps or vps can't be null with h265")
@@ -96,7 +70,7 @@ class RtmpSender(
     }
   }
 
-  fun setAudioInfo(sampleRate: Int, isStereo: Boolean) {
+  override fun setAudioInfo(sampleRate: Int, isStereo: Boolean) {
     audioPacket = when (commandsManager.audioCodec) {
       AudioCodec.G711 -> G711Packet().apply { sendAudioInfo() }
       AudioCodec.AAC -> AacPacket().apply { sendAudioInfo(sampleRate, isStereo) }
@@ -106,82 +80,54 @@ class RtmpSender(
     }
   }
 
-  fun sendMediaFrame(mediaFrame: MediaFrame) {
-    if (running && !queue.trySend(mediaFrame)) {
-      when (mediaFrame.type) {
-        MediaFrame.Type.VIDEO -> {
-          Log.i(TAG, "Video frame discarded")
-          droppedVideoFrames++
-        }
-        MediaFrame.Type.AUDIO -> {
-          Log.i(TAG, "Audio frame discarded")
-          droppedAudioFrames++
-        }
+  override suspend fun onRun() = withContext(Dispatchers.IO) {
+    var bytesSend = 0L
+    val bitrateTask = async {
+      while (scope.isActive && running) {
+        //bytes to bits
+        bitrateManager.calculateBitrate(bytesSend * 8)
+        bytesSend = 0
+        delay(timeMillis = 1000)
       }
     }
-  }
-
-  fun start() {
-    bitrateManager.reset()
-    queue.clear()
-    running = true
-    job = scope.launch {
-      var bytesSend = 0L
-      val bitrateTask = async {
-        while (scope.isActive && running) {
-          //bytes to bits
-          bitrateManager.calculateBitrate(bytesSend * 8)
-          bytesSend = 0
-          delay(timeMillis = 1000)
-        }
-      }
-      while (scope.isActive && running) {
-        val error = runCatching {
-          val mediaFrame = runInterruptible { queue.poll(1, TimeUnit.SECONDS) }
-          getFlvPacket(mediaFrame) { flvPacket ->
-            var size = 0
-            if (flvPacket.type == FlvType.VIDEO) {
-              videoFramesSent++
-              socket?.let { socket ->
-                size = commandsManager.sendVideoPacket(flvPacket, socket)
-                if (isEnableLogs) {
-                  Log.i(TAG, "wrote Video packet, size $size")
-                }
-              }
-            } else {
-              audioFramesSent++
-              socket?.let { socket ->
-                size = commandsManager.sendAudioPacket(flvPacket, socket)
-                if (isEnableLogs) {
-                  Log.i(TAG, "wrote Audio packet, size $size")
-                }
+    while (scope.isActive && running) {
+      val error = runCatching {
+        val mediaFrame = runInterruptible { queue.poll(1, TimeUnit.SECONDS) }
+        getFlvPacket(mediaFrame) { flvPacket ->
+          var size = 0
+          if (flvPacket.type == FlvType.VIDEO) {
+            videoFramesSent++
+            socket?.let { socket ->
+              size = commandsManager.sendVideoPacket(flvPacket, socket)
+              if (isEnableLogs) {
+                Log.i(TAG, "wrote Video packet, size $size")
               }
             }
-            bytesSend += size
+          } else {
+            audioFramesSent++
+            socket?.let { socket ->
+              size = commandsManager.sendAudioPacket(flvPacket, socket)
+              if (isEnableLogs) {
+                Log.i(TAG, "wrote Audio packet, size $size")
+              }
+            }
           }
-        }.exceptionOrNull()
-        if (error != null) {
-          onMainThread {
-            connectChecker.onConnectionFailed("Error send packet, ${error.validMessage()}")
-          }
-          Log.e(TAG, "send error: ", error)
-          return@launch
+          bytesSend += size
         }
+      }.exceptionOrNull()
+      if (error != null) {
+        onMainThread {
+          connectChecker.onConnectionFailed("Error send packet, ${error.validMessage()}")
+        }
+        Log.e(TAG, "send error: ", error)
+        return@withContext
       }
     }
   }
 
-  suspend fun stop(clear: Boolean = true) {
-    running = false
+  override suspend fun stopImp(clear: Boolean) {
     audioPacket.reset(clear)
     videoPacket.reset(clear)
-    resetSentAudioFrames()
-    resetSentVideoFrames()
-    resetDroppedAudioFrames()
-    resetDroppedVideoFrames()
-    job?.cancelAndJoin()
-    job = null
-    queue.clear()
   }
 
   private suspend fun getFlvPacket(mediaFrame: MediaFrame?, callback: suspend (FlvPacket) -> Unit) {
@@ -199,66 +145,4 @@ class RtmpSender(
       }
     }
   }
-
-  @Throws(IllegalArgumentException::class)
-  fun hasCongestion(percentUsed: Float = 20f): Boolean {
-    if (percentUsed < 0 || percentUsed > 100) throw IllegalArgumentException("the value must be in range 0 to 100")
-    val size = queue.size.toFloat()
-    val remaining = queue.remainingCapacity().toFloat()
-    val capacity = size + remaining
-    return size >= capacity * (percentUsed / 100f)
-  }
-
-  fun resizeCache(newSize: Int) {
-    if (newSize < queue.size - queue.remainingCapacity()) {
-      throw RuntimeException("Can't fit current cache inside new cache size")
-    }
-    val tempQueue: BlockingQueue<MediaFrame> = LinkedBlockingQueue(newSize)
-    queue.drainTo(tempQueue)
-    queue = tempQueue
-  }
-
-  fun getCacheSize(): Int {
-    return cacheSize
-  }
-
-  fun getItemsInCache(): Int = queue.size
-
-  fun clearCache() {
-    queue.clear()
-  }
-
-  fun getSentAudioFrames(): Long {
-    return audioFramesSent
-  }
-
-  fun getSentVideoFrames(): Long {
-    return videoFramesSent
-  }
-
-  fun resetSentAudioFrames() {
-    audioFramesSent = 0
-  }
-
-  fun resetSentVideoFrames() {
-    videoFramesSent = 0
-  }
-
-  fun resetDroppedAudioFrames() {
-    droppedAudioFrames = 0
-  }
-
-  fun resetDroppedVideoFrames() {
-    droppedVideoFrames = 0
-  }
-
-  fun setLogs(enable: Boolean) {
-    isEnableLogs = enable
-  }
-
-  fun setBitrateExponentialFactor(factor: Float) {
-    bitrateManager.exponentialFactor = factor
-  }
-
-  fun getBitrateExponentialFactor() = bitrateManager.exponentialFactor
 }
