+++ rtsp/src/main/java/com/pedro/rtsp/rtsp/commands/CommandsManager.kt
@@ -49,9 +49,9 @@ open class CommandsManager {
     private set
   var path: String? = null
     private set
-  var sps: ByteArray? = null
+  var sps: ByteBuffer? = null
     private set
-  var pps: ByteArray? = null
+  var pps: ByteBuffer? = null
     private set
   private var cSeq = 0
   private var sessionId: String? = null
@@ -71,7 +71,7 @@ open class CommandsManager {
   val videoServerPorts = intArrayOf(5006, 5007)
 
   //For H265
-  var vps: ByteArray? = null
+  var vps: ByteBuffer? = null
     private set
 
   //For auth
@@ -100,9 +100,9 @@ open class CommandsManager {
   }
 
   fun setVideoInfo(sps: ByteBuffer, pps: ByteBuffer?, vps: ByteBuffer?) {
-    this.sps = sps.getData()
-    this.pps = pps?.getData()
-    this.vps = vps?.getData()
+    this.sps = sps
+    this.pps = pps
+    this.vps = vps
   }
 
   fun setAudioInfo(sampleRate: Int, isStereo: Boolean) {
@@ -134,11 +134,11 @@ open class CommandsManager {
   }
 
   private val spsString: String
-    get() = sps?.encodeToString() ?: ""
+    get() = sps?.getData()?.encodeToString() ?: ""
   private val ppsString: String
-    get() = pps?.encodeToString() ?: ""
+    get() = pps?.getData()?.encodeToString() ?: ""
   private val vpsString: String
-    get() = vps?.encodeToString() ?: ""
+    get() = vps?.getData()?.encodeToString() ?: ""
 
   private fun addHeaders(): String {
     return "CSeq: ${++cSeq}\r\n" +
